+++ NEWS.md
@@ -5,6 +5,7 @@
 - [Add separate arguments to select release and ppp version](https://github.com/PIP-Technical-Team/pipapi/issues/265)
 
 ## Enhancements
+- [Disable popshare option for pip_grp and aggregate distributions](https://github.com/PIP-Technical-Team/pipapi/pull/274)
 
 # pipapi 1.1.0